# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jio Bihar Diwas API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RechargeRequestInput {
  /** 10-digit mobile number */
  mobileNumber: string;
  /** Referral code or referrer mobile number (optional) */
  referredBy?: string;
}

export interface RechargeRequestResult {
  id: number;
  mobileNumber: string;
  submittedAt: string;
  message: string;
}

export interface AdminVerifyInput {
  password: string;
}

export interface AdminVerifyResult {
  success: boolean;
  token: string;
}

export interface AdminRechargeRequest {
  id: number;
  mobileNumber: string;
  referredBy?: string | null;
  submittedAt: string;
}

export interface AdminRechargeRequestList {
  requests: AdminRechargeRequest[];
  total: number;
}

export interface ErrorResponse {
  error: string;
}

export type AdminGetRechargeRequestsParams = {
  password: string;
};