# lib/api-zod/src/generated/types/adminRechargeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jio Bihar Diwas API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminRechargeRequest {
  id: number;
  mobileNumber: string;
  referredBy?: string | null;
  submittedAt: Date;
}