# lib/api-zod/src/generated/types/adminRechargeRequestList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jio Bihar Diwas API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminRechargeRequest } from "./adminRechargeRequest";

export interface AdminRechargeRequestList {
  requests: AdminRechargeRequest[];
  total: number;
}