# lib/api-zod/src/generated/types/rechargeRequestInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jio Bihar Diwas API
 * OpenAPI spec version: 0.1.0
 */

export interface RechargeRequestInput {
  /** 10-digit mobile number */
  mobileNumber: string;
  /** Referral code or referrer mobile number (optional) */
  referredBy?: string;
}